# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface UploadUrlRequest {
  /**
   * Original file name.
   * @minLength 1
   */
  name: string;
  /**
   * File size in bytes.
   * @minimum 1
   */
  size: number;
  /**
   * MIME type of the file (e.g. `image/jpeg`).
   * @minLength 1
   */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  title: string;
  bio: string;
  photoUrl?: string | null;
  email?: string | null;
  location?: string | null;
  githubUrl?: string | null;
  linkedinUrl?: string | null;
  cvUrl?: string | null;
  skills: string[];
  updatedAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  title?: string;
  bio?: string;
  photoUrl?: string | null;
  email?: string | null;
  location?: string | null;
  githubUrl?: string | null;
  linkedinUrl?: string | null;
  cvUrl?: string | null;
  skills?: string[];
}

export interface Project {
  id: number;
  title: string;
  description: string;
  longDescription?: string | null;
  imageUrl?: string | null;
  technologies: string[];
  category: string;
  githubUrl?: string | null;
  demoUrl?: string | null;
  featured: boolean;
  order: number;
  createdAt: string;
}

export interface CreateProjectBody {
  title: string;
  description: string;
  longDescription?: string | null;
  imageUrl?: string | null;
  technologies: string[];
  category: string;
  githubUrl?: string | null;
  demoUrl?: string | null;
  featured?: boolean;
  order?: number;
}

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  longDescription?: string | null;
  imageUrl?: string | null;
  technologies?: string[];
  category?: string;
  githubUrl?: string | null;
  demoUrl?: string | null;
  featured?: boolean;
  order?: number;
}

export type TimelineEntryType =
  (typeof TimelineEntryType)[keyof typeof TimelineEntryType];

export const TimelineEntryType = {
  education: "education",
  work: "work",
  achievement: "achievement",
} as const;

export interface TimelineEntry {
  id: number;
  title: string;
  organization: string;
  description: string;
  startDate: string;
  endDate?: string | null;
  current: boolean;
  type: TimelineEntryType;
  order: number;
}

export type CreateTimelineEntryBodyType =
  (typeof CreateTimelineEntryBodyType)[keyof typeof CreateTimelineEntryBodyType];

export const CreateTimelineEntryBodyType = {
  education: "education",
  work: "work",
  achievement: "achievement",
} as const;

export interface CreateTimelineEntryBody {
  title: string;
  organization: string;
  description: string;
  startDate: string;
  endDate?: string | null;
  current: boolean;
  type: CreateTimelineEntryBodyType;
  order?: number;
}

export type UpdateTimelineEntryBodyType =
  (typeof UpdateTimelineEntryBodyType)[keyof typeof UpdateTimelineEntryBodyType];

export const UpdateTimelineEntryBodyType = {
  education: "education",
  work: "work",
  achievement: "achievement",
} as const;

export interface UpdateTimelineEntryBody {
  title?: string;
  organization?: string;
  description?: string;
  startDate?: string;
  endDate?: string | null;
  current?: boolean;
  type?: UpdateTimelineEntryBodyType;
  order?: number;
}

export interface Message {
  id: number;
  name: string;
  email: string;
  subject?: string | null;
  content: string;
  read: boolean;
  aiSummary?: string | null;
  aiIntent?: string | null;
  createdAt: string;
}

export interface SendMessageBody {
  name: string;
  email: string;
  subject?: string | null;
  content: string;
}

export interface TrackPageViewBody {
  page: string;
  referrer?: string | null;
}

export type AnalyticsDashboardTopPagesItem = {
  page: string;
  views: number;
};

export type AnalyticsDashboardViewsByDayItem = {
  date: string;
  views: number;
};

export interface AnalyticsDashboard {
  totalViews: number;
  uniqueVisitors: number;
  unreadMessages: number;
  totalMessages: number;
  totalProjects: number;
  topPages: AnalyticsDashboardTopPagesItem[];
  viewsByDay: AnalyticsDashboardViewsByDayItem[];
}

export type AiChatBodyHistoryItemRole =
  (typeof AiChatBodyHistoryItemRole)[keyof typeof AiChatBodyHistoryItemRole];

export const AiChatBodyHistoryItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export type AiChatBodyHistoryItem = {
  role: AiChatBodyHistoryItemRole;
  content: string;
};

export interface AiChatBody {
  message: string;
  history?: AiChatBodyHistoryItem[];
}

export interface AiChatResponse {
  reply: string;
}

export interface ChatMessageBody {
  visitorId: string;
  message: string;
}

export type ChatHistoryResponseItemRole =
  (typeof ChatHistoryResponseItemRole)[keyof typeof ChatHistoryResponseItemRole];

export const ChatHistoryResponseItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export type ChatHistoryResponseItem = {
  role: ChatHistoryResponseItemRole;
  content: string;
  createdAt: string;
};

export type ChatHistoryResponse = ChatHistoryResponseItem[];

export type MessageAnalysisPriority =
  (typeof MessageAnalysisPriority)[keyof typeof MessageAnalysisPriority];

export const MessageAnalysisPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface MessageAnalysis {
  summary: string;
  intent: string;
  suggestedReply: string;
  priority: MessageAnalysisPriority;
}

export interface ProjectSuggestionBody {
  title?: string;
  description?: string;
  category?: string;
  technologies?: string;
}

export interface ProjectSuggestion {
  title: string;
  description: string;
  longDescription: string;
  category: string;
  technologies: string;
}

export type TimelineSuggestionBodyType =
  (typeof TimelineSuggestionBodyType)[keyof typeof TimelineSuggestionBodyType];

export const TimelineSuggestionBodyType = {
  work: "work",
  education: "education",
  achievement: "achievement",
} as const;

export interface TimelineSuggestionBody {
  title?: string;
  organization?: string;
  description?: string;
  type?: TimelineSuggestionBodyType;
}

export type TimelineSuggestionType =
  (typeof TimelineSuggestionType)[keyof typeof TimelineSuggestionType];

export const TimelineSuggestionType = {
  work: "work",
  education: "education",
  achievement: "achievement",
} as const;

export interface TimelineSuggestion {
  title: string;
  organization: string;
  description: string;
  type: TimelineSuggestionType;
}

export interface SecurityAlert {
  id: number;
  attemptedEmail: string;
  attemptedPassword: string;
  ipAddress?: string | null;
  userAgent?: string | null;
  createdAt: string;
}

export interface LogAttemptBody {
  attemptedEmail: string;
  attemptedPassword: string;
}

export type ListProjectsParams = {
  featured?: boolean;
  technology?: string;
  category?: string;
};

export type ListMessagesParams = {
  read?: boolean;
};

export type TrackPageView200 = {
  ok: boolean;
};