# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: technology
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /timeline:
    get:
      operationId: listTimeline
      tags: [timeline]
      summary: List all timeline/experience entries
      responses:
        "200":
          description: Timeline entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEntry"
    post:
      operationId: createTimelineEntry
      tags: [timeline]
      summary: Create a new timeline entry (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimelineEntryBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEntry"
  /timeline/{id}:
    put:
      operationId: updateTimelineEntry
      tags: [timeline]
      summary: Update a timeline entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimelineEntryBody"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEntry"
    delete:
      operationId: deleteTimelineEntry
      tags: [timeline]
      summary: Delete a timeline entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List all messages (admin only)
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a contact message (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}/read:
    patch:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark a message as read (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /chat/history/{visitorId}:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for a visitor
      parameters:
        - name: visitorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the visitor chat widget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: AI reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        technologies:
          type: array
          items:
            type: string
        category:
          type: string
        githubUrl:
          type: string
          nullable: true
        demoUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - technologies
        - category
        - featured
        - order
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        technologies:
          type: array
          items:
            type: string
        category:
          type: string
        githubUrl:
          type: string
          nullable: true
        demoUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        order:
          type: integer
      required:
        - title
        - description
        - technologies
        - category
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        technologies:
          type: array
          items:
            type: string
        category:
          type: string
        githubUrl:
          type: string
          nullable: true
        demoUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        order:
          type: integer
    TimelineEntry:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        organization:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        type:
          type: string
          enum: [education, work, achievement]
        order:
          type: integer
      required:
        - id
        - title
        - organization
        - description
        - startDate
        - current
        - type
        - order
    CreateTimelineEntryBody:
      type: object
      properties:
        title:
          type: string
        organization:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        type:
          type: string
          enum: [education, work, achievement]
        order:
          type: integer
      required:
        - title
        - organization
        - description
        - startDate
        - current
        - type
    UpdateTimelineEntryBody:
      type: object
      properties:
        title:
          type: string
        organization:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        type:
          type: string
          enum: [education, work, achievement]
        order:
          type: integer
    Message:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subject:
          type: string
          nullable: true
        content:
          type: string
        read:
          type: boolean
        aiSummary:
          type: string
          nullable: true
        aiIntent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - content
        - read
        - createdAt
    SendMessageBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
          nullable: true
        content:
          type: string
      required:
        - name
        - email
        - content
...
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
    ChatMessageBody:
      type: object
      properties:
        visitorId:
          type: string
        message:
          type: string
      required:
        - visitorId
        - message
    ChatHistoryResponse:
      type: array
      items:
        type: object
        properties:
          role:
            type: string
            enum: [user, assistant]
          content:
            type: string
          createdAt:
            type: string
        required:
          - role
          - content
          - createdAt
...